Split modifiers and map update #2
